+++ packages/plugin-react/src/config/base.ts
@@ -33,19 +33,21 @@ const getBaseConfig = (chain: WebpackChain) => {
     .set('react-router-dom', loadModule('react-router-dom'))
   chain.module
     .rule('images')
-    .test(/\.(jpe?g|png|woff|woff2|eot|ttf|svg)(\?[a-z0-9=.]+)?$/)
+    .test(/\.(jpe?g|png|woff|woff2|eot|ttf|svg|gif)(\?[a-z0-9=.]+)?$/)
     .use('url-loader')
     .loader(loadModule('url-loader'))
     .options({
       limit: 10000,
-      name: 'static/[name].[hash:8].[ext]',
+      name: '[name].[hash:8].[ext]',
       // require 图片的时候不用加 .default
       esModule: false,
       fallback: {
         loader: loadModule('file-loader'),
         options: {
-          name: 'static/[name].[hash:8].[ext]',
-          esModule: false
+          publicPath: '/client/images',
+          name: '[name].[hash:8].[ext]',
+          esModule: false,
+          outputPath: 'images'
         }
       }
     })