+++ packages/plugin-vue/src/config/server.ts
@@ -1,6 +1,6 @@
 import { join } from 'path'
 import * as webpack from 'webpack'
-import { loadConfig, getLocalNodeModules, getVuexStoreFilePath, nodeExternals } from 'ssr-server-utils'
+import { loadConfig, getLocalNodeModules, nodeExternals } from 'ssr-server-utils'
 import * as WebpackChain from 'webpack-chain'
 import { getBaseConfig } from './base'
 
@@ -36,8 +36,7 @@ const getServerWebpack = (chain: WebpackChain) => {
   })
 
   chain.plugin('define').use(webpack.DefinePlugin, [{
-    __isBrowser__: false,
-    vuexStoreFilePath: JSON.stringify(getVuexStoreFilePath())
+    __isBrowser__: false
   }])
 
   chainServerConfig(chain) // 合并用户自定义配置