+++ packages/server-utils/src/webpack/setStyle.ts
@@ -3,11 +3,12 @@ import { StyleOptions } from 'ssr-types'
 import { loadConfig } from '../loadConfig'
 
 const setStyle = (isDev: boolean, chain: WebpackChain, reg: RegExp, options: StyleOptions, isReact?: boolean) => {
+  const { include, exclude, modules, importLoaders, loader } = options
+  const { css } = loadConfig()
   const MiniCssExtractPlugin = require('mini-css-extract-plugin')
   const loadModule = require.resolve
-  const { css } = loadConfig()
-  const postCssPlugins = css?.().loaderOptions?.postcss?.plugins ?? []
-  const { include, exclude, modules, importLoaders, loader } = options
+
+  const userCssloaderOptions = css?.().loaderOptions?.cssOptions ?? {}
   const cssloaderOptions = {
     importLoaders: importLoaders,
     modules: modules
@@ -16,6 +17,24 @@ const setStyle = (isDev: boolean, chain: WebpackChain, reg: RegExp, options: Sty
     // @ts-expect-error
     cssloaderOptions.getCSSModuleLocalIdent = require('react-dev-utils/getCSSModuleLocalIdent')
   }
+
+  Object.assign(cssloaderOptions, userCssloaderOptions)
+
+  const postCssPlugins = css?.().loaderOptions?.postcss?.plugins ?? [] // 用户自定义 postcss 插件
+  const postCssOptions = Object.assign({
+    ident: 'postcss',
+    plugins: () => [
+      require('postcss-flexbugs-fixes'),
+      require('postcss-discard-comments'),
+      require('postcss-preset-env')({
+        autoprefixer: {
+          flexbox: 'no-2009'
+        },
+        stage: 3
+      })
+    ].concat(postCssPlugins)
+  }, css?.().loaderOptions?.postcss?.options ?? {}) // 合并用户自定义 postcss options
+
   chain.module
     .rule(options.rule)
     .test(reg)
@@ -39,19 +58,7 @@ const setStyle = (isDev: boolean, chain: WebpackChain, reg: RegExp, options: Sty
     .end()
     .use('postcss-loader')
     .loader(loadModule('postcss-loader'))
-    .options({
-      ident: 'postcss',
-      plugins: () => [
-        require('postcss-flexbugs-fixes'),
-        require('postcss-discard-comments'),
-        require('postcss-preset-env')({
-          autoprefixer: {
-            flexbox: 'no-2009'
-          },
-          stage: 3
-        })
-      ].concat(postCssPlugins)
-    })
+    .options(postCssOptions)
     .end()
     .when(Boolean(loader), rule => {
       loader && rule.use(loader)