+++ scripts/scripts.js
@@ -6,10 +6,7 @@ const options = {
   stdio: 'inherit'
 }
 
-const linkPackage = ['@midwayjs/decorator', '@midwayjs/web']
-
-linkPackage.push('react')
-linkPackage.push('react-dom')
+const linkPackage = []
 
 if (argv.bootstrap) {
   let shell = 'npx concurrently "yarn"'
@@ -23,6 +20,11 @@ if (argv.bootstrap) {
   execSync('yarn build:only', options)
 }
 
+const excludePackage = ['.DS_Store', 'plugin-midway', 'plugin-nestjs']
+const packages = fs.readdirSync('./packages').filter(name => {
+  return !excludePackage.includes(name)
+})
+
 if (argv.clean) {
   let shell = 'rm -rf node_modules yarn.lock **/**/cjs **/**/esm **/**/yarn.lock **/**/package-lock.json packages/**/node_modules'
   if (argv.deep) {
@@ -32,7 +34,10 @@ if (argv.clean) {
 }
 
 if (argv.link) {
-  const packages = fs.readdirSync('./packages')
+  if (argv.react) {
+    linkPackage.push('react')
+    linkPackage.push('react-dom')
+  }
   let shell = 'npx concurrently'
   linkPackage.forEach(item => {
     shell += ` "cd node_modules/${item} && yarn link" ` // link react-dom 防止出现多个react实例
@@ -45,11 +50,9 @@ if (argv.link) {
   }
 
   packages.forEach(item => {
-    if (item !== '.DS_Store') {
-      shell += ` "cd packages/${item} && yarn link" ` // link packages 下面所有的包
-    }
+    shell += ` "cd packages/${item} && yarn link" ` // link packages 下面所有的包
   })
-  const linkedPackage = packages.filter(item => item !== '.DS_Store').map(item => item === 'cli' ? 'ssr' : 'ssr-' + item)
+  const linkedPackage = packages.map(item => item === 'cli' ? 'ssr' : 'ssr-' + item)
     .concat(linkPackage).join(' ')
 
   const examples = fs.readdirSync('./example')
@@ -66,7 +69,6 @@ if (argv.unlink) {
     linkPackage.push('react-dom')
   }
 
-  const packages = fs.readdirSync('./packages')
   let shell = 'npx concurrently'
   linkPackage.forEach(item => {
     shell += ` "cd node_modules/${item} && yarn unlink" ` // link react-dom 防止出现多个react实例
@@ -76,17 +78,18 @@ if (argv.unlink) {
     execSync(shell, options)
   }
   packages.forEach(item => {
-    if (item !== '.DS_Store') {
-      shell += ` "cd packages/${item} && yarn unlink" `
-    }
+    shell += ` "cd packages/${item} && yarn unlink" `
   })
   execSync(shell, options)
 }
 if (argv.publishDoc) {
-  const packages = fs.readdirSync('./packages')
   packages.forEach(item => {
-    if (item !== '.DS_Store') {
-      execSync(`cp README.md packages/${item}`, options)
-    }
+    execSync(`cp README.md packages/${item}`, options)
   })
+}
+
+if (argv.sync) {
+  const shell = 'cnpm sync '
+  const linkedPackage = fs.readdirSync('./packages').map(item => item !== '.DS_Store' && item === 'cli' ? 'ssr' : 'ssr-' + item).join(' ')
+  execSync(shell + linkedPackage, options)
 }