+++ packages/plugin-react/CHANGELOG.md
@@ -3,6 +3,116 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.25](https://github.com/ykfe/ssr/compare/v5.4.24...v5.4.25) (2021-03-29)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.24](https://github.com/ykfe/ssr/compare/v5.4.22...v5.4.24) (2021-03-26)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.23](https://github.com/ykfe/ssr/compare/v5.4.22...v5.4.23) (2021-03-25)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.22](https://github.com/ykfe/ssr/compare/v5.4.21...v5.4.22) (2021-03-22)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.20](https://github.com/ykfe/ssr/compare/v5.4.19...v5.4.20) (2021-03-20)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.17](https://github.com/ykfe/ssr/compare/v5.4.16...v5.4.17) (2021-03-20)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.16](https://github.com/ykfe/ssr/compare/v5.4.15...v5.4.16) (2021-03-20)
+
+
+### Bug Fixes
+
+* dynamic css chunk ([bc3fb86](https://github.com/ykfe/ssr/commit/bc3fb86ab11ec27a9919e8d0d9a953f26ab4d851))
+
+
+
+
+
+## [5.4.15](https://github.com/ykfe/ssr/compare/v5.4.14...v5.4.15) (2021-03-20)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.14](https://github.com/ykfe/ssr/compare/v5.4.13...v5.4.14) (2021-03-19)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.13](https://github.com/ykfe/ssr/compare/v5.4.12...v5.4.13) (2021-03-19)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.11](https://github.com/ykfe/ssr/compare/v5.4.10...v5.4.11) (2021-03-18)
+
+
+### Bug Fixes
+
+* file-loader images ([a7f799e](https://github.com/ykfe/ssr/commit/a7f799ee3ff2ccd888c214e4c55216b0ae89c777))
+
+
+
+
+
+## [5.4.10](https://github.com/ykfe/ssr/compare/v5.4.9...v5.4.10) (2021-03-15)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
+## [5.4.9](https://github.com/ykfe/ssr/compare/v5.4.8...v5.4.9) (2021-03-14)
+
+**Note:** Version bump only for package ssr-plugin-react
+
+
+
+
+
 ## [5.4.8](https://github.com/ykfe/ssr/compare/v5.4.7...v5.4.8) (2021-03-13)
 
 **Note:** Version bump only for package ssr-plugin-react